MDL-44846 Atto: Make the color plugins reliable.
The problem is that in IE, the browser implements color commands with font tags (<font color=...>).

We specifically strip font tags to avoid abuse when pasting from word (the world has seen enough of Tahoma).

This solution instead uses Rangy to set inline styles on the selection. This works in all browsers nicely.

I also updated the toggleInlineSelectionClass function to use the same rangy function. This solves problems with
whitespace squashing in ie.