MDL-44844 core: Fix custom menu navigation
Atto introduced a rule to hide all Nodes with the yui3-menu-hidden class
but this rule is too general and applies to non-atto content too.

As a result, when loading the menu and during the XY calculation, the
native Element.getBoundingClientRect() function was returning a bounding
box size of 0 because the Node was not visible and the node was being
hidden from view when it was later made visible.