MDL-45515 atto/plugins: Added border style and BG colour to table
This commit adds the option for borders to be added and
styled, background colours to be set, and width to be set for tables
in the Atto HTML Editor. For those that desire to restrict
use of this, the plugin settings allow the admins to restrict
how much freedom their users have with borders.