MDL-76866 editor_atto: Remove old install script
This install script has not been required for a number of years because
we require the previous LTS for an LTS+1 upgrade. Once Atto was
part of core after an LTS release, it was safe to remove this file.

This is now preventing Tiny from becoming default for new installs in
some situations.